refactor: redundant statements
